:recycle: `useCompleteMenu`でCompleteMenuを表示させるようにする
