+++ web/src/features/editor/components/LetStatementComponent/index.tsx
@@ -1,9 +1,9 @@
 import {
-  CompleteOption,
   EditableNodeComponent,
   StatementWrapper,
   ValidateFn,
 } from '@editor/components';
+import { CompleteOption } from '@editor/hooks/useCompleteMenu';
 import { useInsertNode, useInsertStatement, useStatement } from '@editor/store';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import { Button, Stack, styled } from '@mui/material';
